+++ cloud-siem-20220616/src/Models/ListDisposeStrategyResponseBody/data.php
@@ -5,11 +5,17 @@
 namespace AlibabaCloud\SDK\Cloudsiem\V20220616\Models\ListDisposeStrategyResponseBody;
 
 use AlibabaCloud\Dara\Model;
+use AlibabaCloud\SDK\Cloudsiem\V20220616\Models\ListDisposeStrategyResponseBody\data\groups;
 use AlibabaCloud\SDK\Cloudsiem\V20220616\Models\ListDisposeStrategyResponseBody\data\pageInfo;
 use AlibabaCloud\SDK\Cloudsiem\V20220616\Models\ListDisposeStrategyResponseBody\data\responseData;
 
 class data extends Model
 {
+    /**
+     * @var groups[]
+     */
+    public $groups;
+
     /**
      * @var pageInfo
      */
@@ -20,12 +26,16 @@ class data extends Model
      */
     public $responseData;
     protected $_name = [
+        'groups' => 'Groups',
         'pageInfo' => 'PageInfo',
         'responseData' => 'ResponseData',
     ];
 
     public function validate()
     {
+        if (\is_array($this->groups)) {
+            Model::validateArray($this->groups);
+        }
         if (null !== $this->pageInfo) {
             $this->pageInfo->validate();
         }
@@ -38,6 +48,17 @@ class data extends Model
     public function toArray($noStream = false)
     {
         $res = [];
+        if (null !== $this->groups) {
+            if (\is_array($this->groups)) {
+                $res['Groups'] = [];
+                $n1 = 0;
+                foreach ($this->groups as $item1) {
+                    $res['Groups'][$n1] = null !== $item1 ? $item1->toArray($noStream) : $item1;
+                    ++$n1;
+                }
+            }
+        }
+
         if (null !== $this->pageInfo) {
             $res['PageInfo'] = null !== $this->pageInfo ? $this->pageInfo->toArray($noStream) : $this->pageInfo;
         }
@@ -64,6 +85,17 @@ class data extends Model
     public static function fromMap($map = [])
     {
         $model = new self();
+        if (isset($map['Groups'])) {
+            if (!empty($map['Groups'])) {
+                $model->groups = [];
+                $n1 = 0;
+                foreach ($map['Groups'] as $item1) {
+                    $model->groups[$n1] = groups::fromMap($item1);
+                    ++$n1;
+                }
+            }
+        }
+
         if (isset($map['PageInfo'])) {
             $model->pageInfo = pageInfo::fromMap($map['PageInfo']);
         }